refactor: extract state-serialization & game-summary cluster from state.py (#1271)
Pull the frontend/StatsService serialization entry points plus the
end-of-game summary / live performance dict builders out of the
state.py God-Object into a new StateSerializationMixin
(game/state_serialization.py).

Moved (behavior-preserving, byte-identical code move):
- get_state / get_reveal_players_state (GameStateSerializer delegators)
- set_stats_service / get_song_difficulty (StatsService wiring + lookup)
- finalize_game (end-of-game summary dict, Story 14.4)
- _calculate_current_avg / get_game_performance (live performance compare)

GameStateSerializer is now imported lazily inside the two view delegators
(noqa: PLC0415) instead of at module top, removing the state.py top-level
import and matching the lazy-import discipline of the sibling state_*.py
mixins. Public API + every caller / test unchanged.

state.py: 1955 -> 1818 lines (-137).
